mqttclient/requirements.txt
#ruamel.yaml  installed seperately due to error
paho-mqtt>=1.6.1
